Move CC and iDeal payments to payments helper
